Fix the scaledOffset alias type
ScaledOffset was an alias to int64_t,
which resulted in build errors, because
the mako template produced floating-point value.
Now ScaledOffset is an alias to a double type.

+++ scripts/writesensor.mako.cpp
@@ -70,7 +70,7 @@ extern const IdInfoMap sensors = {
 %>
         ${entityID},${instance},${sensorType},"${path}","${sensorInterface}",
         ${readingType},${multiplier},${offsetB},${bExp},
-        ${offsetB * int(pow(10,bExp))}, ${rExp}, ${hasScale},${scale},"${unit}",
+        ${offsetB * pow(10,bExp)}, ${rExp}, ${hasScale},${scale},"${unit}",
         ${updateFunc},${getFunc},Mutability(${mutability}),${sensorNameFunc},{
     % for interface,properties in interfaces.items():
             {"${interface}",{
